+++ src/OpenTelemetry.Exporter.ZPages/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Renamed extension method from `UseZPagesExporter` to `AddZPagesExporter`
   ([#1066](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1066))
-* Changed `ZPagesExporter` to use `BatchExportActivityProcessor` by default
+* Changed `ZPagesExporter` to use `ZPagesProcessor` by default
   ([#1108](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1108))
 
 ## 0.4.0-beta.2